# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Gym CRM API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/members/{member_id}/checkups": {
      "get": {
        "tags": [
          "checkups"
        ],
        "summary": "List Checkups",
        "operationId": "list_checkups_api_members__member_id__checkups_get",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CheckUpOut"
                  },
                  "title": "Response List Checkups Api Members  Member Id  Checkups Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "checkups"
        ],
        "summary": "Create Checkup",
        "operationId": "create_checkup_api_members__member_id__checkups_post",
        "parameters": [
          {
            "name": "member_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Member Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckUpIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUpOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkups/{checkup_id}": {
      "patch": {
        "tags": [
          "checkups"
        ],
        "summary": "Update Checkup",
        "operationId": "update_checkup_api_checkups__checkup_id__patch",
        "parameters": [
          {
            "name": "checkup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checkup Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckUpUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUpOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "checkups"
        ],
        "summary": "Delete Checkup",
        "operationId": "delete_checkup_api_checkups__checkup_id__delete",
        "parameters": [
          {
            "name": "checkup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checkup Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkups/{checkup_id}/photos": {
      "post": {
        "tags": [
          "checkups"
        ],
        "summary": "Upload Photo",
        "operationId": "upload_photo_api_checkups__checkup_id__photos_post",
        "parameters": [
          {
            "name": "checkup_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Checkup Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_api_checkups__checkup_id__photos_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUpOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkup-photos/{photo_id}": {
      "get": {
        "tags": [
          "checkups"
        ],
        "summary": "Get Photo",
        "operationId": "get_photo_api_checkup_photos__photo_id__get",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "checkups"
        ],
        "summary": "Delete Photo",
        "operationId": "delete_photo_api_checkup_photos__photo_id__delete",
        "parameters": [
          {
            "name": "photo_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Photo Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckUpOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/checkups/due": {
      "get": {
        "tags": [
          "checkups"
        ],
        "summary": "Due Checkups",
        "description": "Members with a running membership whose last check-up is older than the gym's\ncheck-up interval (or who have never had one). Longest-waiting first.",
        "operationId": "due_checkups_api_checkups_due_get",
        "parameters": [
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "mine",
                "all"
              ],
              "type": "string",
              "default": "all",
              "title": "Scope"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_DueCheckUp_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_photo_api_checkups__checkup_id__photos_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_photo_api_checkups__checkup_id__photos_post"
      },
...
      "CheckUpIn": {
        "properties": {
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 350.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 260.0,
                "minimum": 50.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 75.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          },
          "muscle_mass_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 200.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Muscle Mass Kg"
          },
          "chest_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chest Cm"
          },
          "waist_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Waist Cm"
          },
          "hips_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hips Cm"
          },
          "arm_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Arm Cm"
          },
          "thigh_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 150.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Thigh Cm"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recorded_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded On"
          }
        },
        "type": "object",
        "title": "CheckUpIn"
      },
      "CheckUpOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "member_id": {
            "type": "string",
            "format": "uuid",
            "title": "Member Id"
          },
          "recorded_on": {
            "type": "string",
            "format": "date",
            "title": "Recorded On"
          },
          "weight_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "bmi": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bmi"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          },
          "muscle_mass_kg": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Muscle Mass Kg"
          },
          "chest_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chest Cm"
          },
          "waist_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waist Cm"
          },
          "hips_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hips Cm"
          },
          "arm_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arm Cm"
          },
          "thigh_cm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Thigh Cm"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recorded_by": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/StaffRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "photos": {
            "items": {
              "$ref": "#/components/schemas/PhotoOut"
            },
            "type": "array",
            "title": "Photos"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "member_id",
          "recorded_on",
          "weight_kg",
          "height_cm",
          "bmi",
          "body_fat_pct",
          "muscle_mass_kg",
          "chest_cm",
          "waist_cm",
          "hips_cm",
          "arm_cm",
          "thigh_cm",
          "notes",
          "recorded_by",
          "photos",
          "created_at"
        ],
        "title": "CheckUpOut"
      },
      "CheckUpUpdate": {
        "properties": {
          "weight_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 350.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Weight Kg"
          },
          "height_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 260.0,
                "minimum": 50.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Height Cm"
          },
          "body_fat_pct": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 75.0,
                "minimum": 2.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Body Fat Pct"
          },
          "muscle_mass_kg": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 200.0,
                "minimum": 5.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Muscle Mass Kg"
          },
          "chest_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Chest Cm"
          },
          "waist_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Waist Cm"
          },
          "hips_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 250.0,
                "minimum": 30.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Hips Cm"
          },
          "arm_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 100.0,
                "minimum": 10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Arm Cm"
          },
          "thigh_cm": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 150.0,
                "minimum": 20.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Thigh Cm"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "recorded_on": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recorded On"
          }
        },
        "type": "object",
        "title": "CheckUpUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_DueCheckUp_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DueCheckUp"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[DueCheckUp]"
      },
...
      "PhotoOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "url",
          "created_at"
        ],
        "title": "PhotoOut"
      },
...
      "StaffRef": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "StaffRef"
      },
...
    }
  }
}